==> Step0_freguesias_latlog.ipynb <==
import pandas as pd
import numpy as np

##Get latitude and longitude of all freguesias of Lisbon:

from geopy.geocoders import Nominatim
geolocator = Nominatim(user_agent="crocs49")
location = geolocator.geocode("Pena,Lisboa ")
print(location.address)
print((location.latitude, location.longitude))



freguesias_df = pd.read_excel(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\GoogleMaps_Places\Freguesias.xlsx")

type(freguesias_df["Freguesia"].iloc[0])

freguesias_df.head(5)

#Treating beginning of dataframe:
for i in range(len(freguesias_df)):
    freguesias_df["Freguesia"].iloc[i] = freguesias_df["Freguesia"].iloc[i].split("nota")
    
for i in range(len(freguesias_df)):
    freguesias_df["Freguesia"].iloc[i] = freguesias_df["Freguesia"].iloc[i][0].replace("[","")

freguesias_df.head(5)

freguesias_df.to_excel(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\GoogleMaps_Places\Freguesias.xlsx")

freguesias_df = pd.read_excel(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\GoogleMaps_Places\Freguesias.xlsx")

freguesias_df.drop(columns = ["Unnamed: 3","Index"], inplace = True)

freguesias_df.head(2)

location = geolocator.geocode(freguesias_df["Freguesia"][1]+","+freguesias_df["Cidade"][1])
print(location.address)
print((location.latitude, location.longitude))

def lat_log(row):
    row["lat"] = np.nan
    row["lon"] = np.nan
    
    for i in range(len(row)):
        location = geolocator.geocode(freguesias_df["Freguesia"][i]+","+freguesias_df["Cidade"][i])
        row["lat"].iloc[i] = location.latitude
        row["lon"].iloc[i] = location.longitude
        
    return row

freguesias_df = lat_log(freguesias_df)

freguesias_df.head(5)

freguesias_df.to_excel(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\GoogleMaps_Places\Freguesias.xlsx")
