==> Step02_enpi_webscrape.ipynb <==
# this is the library that we will use to create break times in order to mimic human behaviour 
import time

# this library will be used for getting your credentials but keeping them hidden in the script 
from getpass import getpass

# these are the libraries we will use for interaction with a webpage
from selenium import webdriver
from selenium.webdriver.common.keys import Keys

# library for interacting with the operating system
import os

# you know pandas it's your best buddy 
import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'
import numpy as np

#Import beautiful soup and requests
from bs4 import BeautifulSoup
import requests

business_dset = pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\google_maps_mycode\health_database.csv")

print(business_dset.shape)
business_dset.head(4)

# driver path

## driver path should be the location of the folder where the driver you downloaded is ## 

driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 

## first we need to initiate the driver - probably the most important part of the code ##
driver = webdriver.Chrome(executable_path = driver_path)

# open the website
driver.get('https://servicosonline.inpi.justica.gov.pt/pesquisas/main/marcas.jsp?lang=PT')

# finding the element that says "Pesquisa Fonética (para Nomes de Estabelecimentos, Insígnias e Logotipos)":

filter_button = driver.find_element_by_id("ext-gen63")
filter_button.click() ## with the method click you will be able to click the element you searched for ##

time.sleep(1.5) ## never forget to put a sleep timer in order to mimic human interaction ##

# finding the element that says "Continuar" and click on it:
continue_button = driver.find_element_by_id("ext-gen107")
continue_button.click() ## with the method click you will be able to click the element you searched for ##
time.sleep(1.5)

#After we  enter the brand search page, we need to put the name of the brand we want to search in the search area and click "Pesquisar" as well:

#First we see what we want to search. in this case it's every company name that we have in our dataframe. For example main_df["name"][6] = MA & CE
search_string = main_df["name"][6]

# here we are going to find the box where we input the email
brand_input = driver.find_element_by_id("ext-comp-1029")
# with the method send_keys() you can send the information from the script to the webpage
brand_input.send_keys(search_string)
time.sleep(0.5)

#We just  need to find and hit the "Pesquisar" button now:

continue_button = driver.find_element_by_id("ext-gen121")
continue_button.click() ## with the method click you will be able to click the element you searched for ##
time.sleep(10)

#After we search, we want the company that has the highest %Identidade. Luckily, the table is already order DESCENDING so it's the first result that we want:

wanted_company = driver.find_element_by_xpath('//*[@id="ext-gen136"]/div[1]/table/tbody/tr/td[2]/div/div/a')
wanted_company.click()



labels = driver.find_elements_by_xpath('//*[@id="ext-gen222"]/div/div/table/tbody/tr/td[1]/div/table/tbody/tr/td[1]')
for label in labels:
    print(label.text)
print("""
space""")
#values = driver.find_elements_by_xpath('//*[@id="ext-gen222"]/div/div/table/tbody/tr/td[1]/div/table/tbody/tr/td[2]')
values = driver.find_elements_by_xpath('//*[@id="ext-gen222"]/div/div/table/tbody/tr/td[1]/div/table/tbody/tr/td[2]')



razao_social = values[15].text
mandatario = values[16].text
morada = values[17].text
codigo_postal = values[18].text

print(razao_social)
print(mandatario)
print(morada)

###Now everything in a function:

def get_razao_social (row) :
    row["ano_fundacao"] = np.nan
    row["razao_social"] = np.nan
    row["mandatario"] = np.nan
    row["morada"] = np.nan
    row["codigo_postal"] = np.nan

    list_uncaptured = []
    # driver path
## driver path should be the location of the folder where the driver you downloaded is ## 

    driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 

## first we need to initiate the driver - probably the most important part of the code ##
    driver = webdriver.Chrome(executable_path = driver_path)
    
    for i in range(len(row)):
    
        # open the website
        driver.get('https://servicosonline.inpi.justica.gov.pt/pesquisas/main/marcas.jsp?lang=PT')
        # finding the element that says "Pesquisa Fonética (para Nomes de Estabelecimentos, Insígnias e Logotipos)":
        time.sleep(3.5)
        filter_button = driver.find_element_by_id("ext-gen63")
        filter_button.click() ## with the method click you will be able to click the element you searched for ##

        time.sleep(1.5) ## never forget to put a sleep timer in order to mimic human interaction ##

        # finding the element that says "Continuar" and click on it:
        continue_button = driver.find_element_by_id("ext-gen107")
        continue_button.click() ## with the method click you will be able to click the element you searched for ##
        time.sleep(1.5)

        #After we  enter the brand search page, we need to put the name of the brand we want to search in the search area and click "Pesquisar" as well:

        #First we see what we want to search. in this case it's every company name that we have in our dataframe. For example main_df["name"][6] = MA & CE
        try: 
            search_string = row["Name"][i]

            # here we are going to find the box where we input the email
            brand_input = driver.find_element_by_id("ext-comp-1029")
            # with the method send_keys() you can send the information from the script to the webpage
            brand_input.send_keys(search_string)
            time.sleep(1)

            #We just  need to find and hit the "Pesquisar" button now:

            continue_button = driver.find_element_by_id("ext-gen121")
            continue_button.click() ## with the method click you will be able to click the element you searched for ##
            time.sleep(13)

            #After we search, we want the company that has the highest %Identidade. Luckily, the table is already order DESCENDING so it's the first result that we want:
            wanted_company = driver.find_element_by_xpath('//*[@id="ext-gen136"]/div[1]/table/tbody/tr/td[2]/div/div/a')
            wanted_company.click()
            time.sleep(2.5)

            #Search all . ATTENTION we need to inspect X-Path to check if the gen number stays the same or changes
            values = driver.find_elements_by_xpath('//*[@id="ext-gen222"]/div/div/table/tbody/tr/td[1]/div/table/tbody/tr/td[2]')
            time.sleep(2.5)
            #Append information in the dataframes
            row["razao_social"].iloc[i] = values[15].text
            row["ano_fundacao"].iloc[i]  = values[0].text
            row["mandatario"].iloc[i] = values[16].text
            row["morada"].iloc[i] = values[17].text
            row["codigo_postal"].iloc[i] = values[18].text
        except:
            print("didn't capture",row["name"][i])
            list_uncaptured += row["name"][i]
            
    return row 

#3Aplying the function to inplace it in the dataframe

razao_social_df = get_razao_social(business_dset)


razao_social_df["mandatario"].isnull().sum()
razao_social_df.head(10)

razao_social_df.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\razao_social_df.csv")

razao_social_df["razao_social"].value_counts
