==> Step03_facebook_webscraping.ipynb <==
# this is the library that we will use to create break times in order to mimic human behaviour 
import time

# this library will be used for getting your credentials but keeping them hidden in the script 
from getpass import getpass

# these are the libraries we will use for interaction with a webpage
from selenium import webdriver
from selenium.webdriver.common.keys import Keys

# library for interacting with the operating system
import os

# you know pandas it's your best buddy 
import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'
import numpy as np

#Import beautiful soup and requests
from bs4 import BeautifulSoup
import requests

import re
import random

#import dataframe:
business_dset = pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\google_maps_mycode\health_database.csv")
print(business_dset.shape)
business_dset.head(10)

# driver path

## driver path should be the location of the folder where the driver you downloaded is ## 

driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 

## first we need to initiate the driver - probably the most important part of the code ##
driver = webdriver.Chrome(executable_path = driver_path)

# open the website
driver.get('https://facebook.com/?ab=0')
time.sleep(2+ random.randint(0,5) + random.random())

# close the information button:

#Check 1 time to see if it pops up or not:

info_button = driver.find_element_by_css_selector("button[class^='_42ft _4jy0 _9xo7']")
info_button.click() ## with the method click you will be able to click the element you searched for ##
time.sleep(1+ random.randint(0,5) + random.random())

#driver.find_elements_by_css_selector("button[class^='_42ft _4jy0 _9xo7']")

#LOG IN:
#Insert username and password:
e_mail = driver.find_element_by_css_selector("input[class^='inputtext']")
e_mail.click()
e_mail.send_keys("goncalodjantunes@gmail.com")
 
time.sleep(1+ random.randint(0,5) + random.random())

password = driver.find_element_by_css_selector("input[class^='inputtext _55r1 _6luy _9']")
password.click()
password.send_keys("GoncaloAntunes29")


time.sleep(3+ random.randint(0,5) + random.random())
#clicking in the start session button:
filter_button = driver.find_element_by_css_selector("button[class^='_42ft _4jy0 _6lth _4jy6 _4jy1']")
filter_button.click()

time.sleep(2+ random.randint(0,5) + random.random())




#Search companies profiles:

search_button = driver.find_element_by_css_selector("label[class^='rq0escxv a8c37x1j a5nuqjux l9j0dhe7 ']")
search_button.click()
time.sleep(1.5 +random.random())
search_button.send_keys("Farmácia Internacional")
search_button.send_keys(Keys.ENTER)
time.sleep(1+ random.randint(0,5) + random.random())



#Select only business pages:
business_page_button = driver.find_elements_by_css_selector("div[data-visualcompletion^='ignore-dynamic']")
time.sleep(1+ random.randint(0,5) + random.random())
print(len(business_page_button))
for business in business_page_button:
    print(business.text,"i")
if len(business_page_button) < 20:
    business_page_button[7].click()

else:
     business_page_button[20].click()
# for business in business_page_button:
#     print(business.text)

#d2edcug0 hpfvmrgz qv66sw1b c1et5uql lr9zc1uh jq4qci2q a3bd9o3v ekzkrbhg oo9gr5id



#select the first business:
time.sleep(3+ random.randint(0,2) + random.random())
wanted_business = driver.find_elements_by_css_selector("span[class^='nc684nl6']")
wanted_business[0].click()


html = driver.page_source

soup = BeautifulSoup(html)

#soup.prettify()

##For business pages:
table = soup.find_all("div",attrs = {"class": "datstx6m cbu4d94t j83agx80"}) 

time.sleep(2+ random.randint(0,5) + random.random())



#j83agx80 l9j0dhe7 k4urcfbm

for element in table:
    print(element.text,"\n")

contacts =re.findall("((\+\d{3} ?|\+\d{2} ?)?)((\d{3} ?){3}|(\d{2} \d{3} \d{4}?))", str(table))
email = re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", str(table))
if len(email) >0:
    email = email[0]
else:
    email = 0


#like = soup.find("div",text=re.compile('pessoas gostam disto'))
for element in table:
    followers = re.findall("disto ?[0-9]+? ?[0-9]+? ?pessoas",element.text)
    if str(followers) ==0:
        following = 0        
    elif len(followers)>0:
        following =  re.findall("\d+",followers[0])
    else:
        following = 0
    website = re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',element.text)
#     if len(following) >  0:
#         following = following[0]
#     else:
#         following = 0
        
    if len(website) > 0:
        website = website[0]
    else:
        website = 0
#print(len(table))
print(contacts)
print(email)
print(followers)
print(following)
print(website)

##For pages:

table2 = soup.find_all("a",attrs = {"class": "oajrlxb2 g5ia77u1 qu0x051f esr5mh6w e9989ue4 r7d6kgcz rq0escxv nhd2j8a9 nc684nl6 p7hjln8o kvgmc6g5 cxmmr5t8 oygrvhab hcukyx3x jb3vyjys rz4wbd8a qt6c0cv9 a8nywdso i1ao9s8h esuyzwwr f1sip0of lzcic4wl gpro0wi8 m9osqain lrazzd5p"}) 
time.sleep(1+ random.randint(0,3) + random.random())
if len(table2)>0:
    table2= table2[0]
else:
    table2=table2
for element in table2:
    followers= element

print(followers)

table3 = soup.find_all("div",attrs = {"class": "rq0escxv l9j0dhe7 du4w35lb j83agx80 cbu4d94t d2edcug0 hpfvmrgz rj1gh0hx buofh1pr g5gj957u p8fzw8mz pcp91wgn iuny7tx3 ipjc6fyt"}) 
contacts1 =re.findall("((\+\d{3} ?|\+\d{2} ?)?)((\d{3} ?){3}|(\d{2} \d{3} \d{4}?))", str(table3))
email1 = re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", str(table3))

for element in table3:

    website1 = re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',element.text)
print(contacts1[0])
print(email1)
print(website1)
        



#Grab email:

table3 = soup.find_all("div",attrs = {"class": "rq0escxv l9j0dhe7 du4w35lb j83agx80 pfnyh3mw i1fnvgqd gs1a9yip owycx6da btwxx1t3 hv4rvrfc dati1w0a discj3wi b5q2rw42 lq239pai mysgfdmx hddg9phg"}) 
#table4 = soup.find_all("div",attrs =  {"class": "rq0escxv l9j0dhe7 du4w35lb j83agx80 cbu4d94t d2edcug0 hpfvmrgz rj1gh0hx buofh1pr g5gj957u p8fzw8mz pcp91wgn iuny7tx3 ipjc6fyt"})

print(13)
contacts1 =re.findall("((\+\d{3} ?|\+\d{2} ?)?)((\d{3} ?){3}|(\d{2} \d{3} \d{4}?))", str(table3))
#row["fb_phone_num"].iloc[i] = contacts1[0]
contacts = contacts1[0]
#row["email"].iloc[i] = re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", str(table3))
emails = re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", str(table3))

print(14)
for element in table3:
    print(element.text)
    website = re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',element.text)
    print(website)

print(website)
print(contacts)
print(emails)

table
for element in table:
    print(re.search('(.*)pessoas', element.text))
    print(element.text)

string = ".Na Shop1One encontras soluções de qualidade +351 21 386 3409 a preços competitivos.22 979 pessoas gostam disto 201 pessoas seguem +351 969 343 291 istohttp://www.Shop1One.com/Enviar mensageminfo@shop1one.com +55 913 541 277"


print(re.findall("disto ?[0-9]+? ?[0-9]+? pessoas",string))


print(re.findall("((\+\d{3} ?|\+\d{2} ?)?)((\d{3} ?){3}|(\d{2} \d{3} \d{4}?))", string))
print(re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',string))


##Full function for facebook wescraping:

###Now everything in a function:

def get_fb_info (row) :
    row["fb_phone_num"] = np.nan
    row["email"] = np.nan
    row["following_fb"] = np.nan
    row["website_fb"] = np.nan


    list_uncaptured = []
    # driver path
## driver path should be the location of the folder where the driver you downloaded is ## 

    driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 

## first we need to initiate the driver - probably the most important part of the code ##
    driver = webdriver.Chrome(executable_path = driver_path)
        
    # open the website

    driver.get('https://facebook.com/?ab=0')
    time.sleep(2+ random.randint(0,5) + random.random())

    # close the information button:
    
    info_button = driver.find_element_by_css_selector("button[class^='_42ft _4jy0 _9xo7']")
    info_button.click() ## with the method click you will be able to click the element you searched for ##
    time.sleep(1+ random.randint(0,5) + random.random()) ## never forget to put a sleep timer in order to mimic human interaction ##

    #LOG IN:
    #Insert username and password:
    e_mail = driver.find_element_by_css_selector("input[class^='inputtext']")
    e_mail.click()
    e_mail.send_keys("nerewoh826@ningame.com")
    time.sleep(1+ random.randint(0,5) + random.random())

    password = driver.find_element_by_css_selector("input[class^='inputtext _55r1 _6luy _9']")
    password.click()
    password.send_keys("JoaoAntunes29")
    time.sleep(2+ random.randint(0,3) + random.random())

    #clicking in the start session button:
    filter_button = driver.find_element_by_css_selector("button[class^='_42ft _4jy0 _6lth _4jy6 _4jy1']")
    filter_button.click()
    time.sleep(1+ random.randint(0,5) + random.random())

    for i in range(len(row)):

    #After we  enter the brand search page, we need to put the name of the brand we want to search in the search area and click "Pesquisar" as well:

    #First we see what we want to search. in this case it's every company name that we have in our dataframe. For example main_df["name"][6] = MA & CE
        try: 
            #Search companies profiles:
            time.sleep(1+ random.randint(0,2) + random.random())
            search_button = driver.find_element_by_css_selector("label[class^='rq0escxv a8c37x1j a5nuqjux l9j0dhe7 ']")
            search_button.click()
            time.sleep(1+ random.randint(0,3) + random.random())
            search_button.send_keys(row["Name"].iloc[i])
            search_button.send_keys(Keys.ENTER)
            time.sleep(2+ random.randint(0,3) + random.random())
            
            print("0")
            #Select only business pages:
            business_page_button = driver.find_elements_by_css_selector("div[data-visualcompletion^='ignore-dynamic']")
            time.sleep(1+ random.randint(0,3) + random.random())
            print(len(business_page_button))
#             for element in business_page_button:
#                 print(element.text)
            if len(business_page_button) < 20:
                print("1")

                business_page_button[6].click()
                
            else:
                print("2")
                business_page_button[19].click()

            #select the first business:
            time.sleep(3+ random.randint(0,5) + random.random())
            print("3")
            wanted_business = driver.find_elements_by_css_selector("span[class^='nc684nl6']")
            wanted_business[0].click()
            time.sleep(3+ random.randint(0,5) + random.random())           
            print("4")
            #After entering the business page we have to do beautiful soup in it to extract the info:
            html = driver.page_source
            time.sleep(1+random.random())
            soup = BeautifulSoup(html)
            print("5")
            time.sleep(2+random.random())

            #In case it is a business page:
            table = soup.find_all("div",attrs = {"class": "datstx6m cbu4d94t j83agx80"}) 
            if len(table)>0:
           #Append information in the dataframes
            #Get phone_num:
                row["fb_phone_num"].iloc[i]= re.findall("((\+\d{3} ?|\+\d{2} ?)?)((\d{3} ?){3}|(\d{2} ?\d{3} ?\d{4}?))", str(table))
                print(8)
                #Get email:    
                row["email"].iloc[i] = re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", str(table))
                print(9)
    #             if(len(email)) >0:
    #                 row["email"].iloc[i] = email[0]
    #                 print(10)
    #             else:
    #                 row["email"].iloc[i] = 0    
    #                 print(11)
    #             #For loop to append inside info:
                time.sleep(1+ random.randint(0,5) + random.random())
                for element in table:
                    print(12)
                    #Get followers number:

                    row["following_fb"].iloc[i] = re.findall("disto ?[0-9]+? ?[0-9]+? ?pessoas",element.text)

                    #Get website link:
                    row["website_fb"].iloc[i] = re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',element.text)
                    
            #In case it is a normal page:    
            else:
                table2 = soup.find_all("a",attrs = {"class": "oajrlxb2 g5ia77u1 qu0x051f esr5mh6w e9989ue4 r7d6kgcz rq0escxv nhd2j8a9 nc684nl6 p7hjln8o kvgmc6g5 cxmmr5t8 oygrvhab hcukyx3x jb3vyjys rz4wbd8a qt6c0cv9 a8nywdso i1ao9s8h esuyzwwr f1sip0of lzcic4wl gpro0wi8 m9osqain lrazzd5p"}) 
                time.sleep(4+ random.randint(0,5) + random.random())
                if len(table2)>0:
                    table2= table2[0]
                else:
                    table2=table2
                for element in table2:
                    row["following_fb"].iloc[i]= element
                    print(12.5)
                #print(followers)

                table3 = soup.find_all("div",attrs = {"class": "rq0escxv l9j0dhe7 du4w35lb j83agx80 cbu4d94t d2edcug0 hpfvmrgz rj1gh0hx buofh1pr g5gj957u p8fzw8mz pcp91wgn iuny7tx3 ipjc6fyt"}) 
                #rq0escxv l9j0dhe7 du4w35lb j83agx80 pfnyh3mw i1fnvgqd gs1a9yip owycx6da btwxx1t3 hv4rvrfc dati1w0a discj3wi b5q2rw42 lq239pai mysgfdmx hddg9phg
                #rq0escxv l9j0dhe7 du4w35lb j83agx80 cbu4d94t d2edcug0 hpfvmrgz rj1gh0hx buofh1pr g5gj957u p8fzw8mz pcp91wgn iuny7tx3 ipjc6fyt
                print(13)
                contacts1 =re.findall("((\+\d{3} ?|\+\d{2} ?)?)((\d{3} ?){3}|(\d{2} \d{3} \d{4}?))", str(table3))
                print(13.5)
                row["fb_phone_num"].iloc[i] = contacts1[0]
                row["email"].iloc[i] = re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", str(table3))
                print(14)
                for element in table3:
                    website = re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',element.text)
                    print(website)
                    row["website_fb"].iloc[i] = re.findall('https?://(?:[-\w.]|(?:%[\da-fA-F]{2}))+',element.text)
                    print(15)


        except:
            print("didn't capture",row["Name"].iloc[i])
            list_uncaptured += row["Name"].iloc[i]

    return row 





#Lets import the 250 rows of our dataframe:

data = pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\razao_social_251_600.csv")
data1 = data[:113]

fb_df = get_fb_info(data1)

fb_df.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\facebook250_362.csv")

fb_df[fb_df["Name"]=="Farmácia Tavares de Matos e Neca Lda"]

fb_df.head(35)
