==> Step04_dbhoovers_webscraping.ipynb <==
##Using Selenium to webscrape dbhoovers website:


# this is the library that we will use to create break times in order to mimic human behaviour 
import time

# this library will be used for getting your credentials but keeping them hidden in the script 
from getpass import getpass

# these are the libraries we will use for interaction with a webpage
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC

# library for interacting with the operating system
import os

# you know pandas it's your best buddy 
import pandas as pd
pd.options.mode.chained_assignment = None  # default='warn'
import numpy as np

#Import beautiful soup and requests
from bs4 import BeautifulSoup
import requests

  
#Import dataframe
data  =pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\facebook_50.csv")
#type(data["razao_social"].iloc[0])
data["razao_social"] =  data["razao_social"].astype(str)
data.shape

from selenium import webdriver

driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 


options = webdriver.ChromeOptions()
#options.add_argument('--headless')
options.add_argument('--no-sandbox')
options.add_argument('--window-size=1420,1080')
options.add_argument('--disable-gpu')
options.add_argument(f'user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.93 Safari/537.36')
options.add_experimental_option("excludeSwitches", ["enable-automation"])
options.add_experimental_option('useAutomationExtension', False)
driver = webdriver.Chrome(options=options, executable_path= driver_path)
driver.get("https://www.dnb.com/solutions/small-business/small-business-strategic-planning.html")



# options.add_argument(f'user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.93 Safari/537.36')

# driver = webdriver.Chrome(executable_path = driver_path )#, options = options)

# driver.get("https://www.dnb.com/solutions/small-business/small-business-strategic-planning.html")
# driver.delete_all_cookies()
 


data.head(5)

##Step 0:

# ##Web-scraping dnb.com (D&B Hoovers)
#     # driver path
driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 

#     ## first we need to initiate the driver - probably the most important part of the code ##
# driver = webdriver.Chrome(executable_path = driver_path)
# options = webdriver.ChromeOptions() 
# options.add_argument("start-maximized")
# options.add_experimental_option("excludeSwitches", ["enable-automation"])
# options.add_experimental_option('useAutomationExtension', False)

##1st:

    #Open the website
driver.get('https://www.dnb.com/business-directory/company-search.html')
    # finding the element that says "Pesquisa Fonética (para Nomes de Estabelecimentos, Insígnias e Logotipos)":
time.sleep(3)
#filter_button = driver.find_element_by_id("query_display")
#filter_button.click() ## with the method click you will be able to click the element you searched for ##


##2nd:

    #Find the top search area

search = driver.find_element_by_xpath('//*[@id="search_button"]')
search.click()




##3rd:

    #Go to the search area and put the string we want:
find = driver.find_element_by_xpath('//*[@id="search_query"]')
time.sleep(1.5)
find.click()
find = driver.find_element_by_xpath('//*[@id="search_query"]')
find.clear()
time.sleep(1.5)
find.send_keys("Gray's Cuisine")
time.sleep(1)
#find.send_keys(Keys.ENTER)



#4th:

    #Click enter:
time.sleep(1)
find.send_keys(Keys.ENTER)

#5th:

    #Click on the first result for the companies list and wait a bit for the page to load:
pretended_company =driver.find_element_by_xpath('//*[@id="company_results"]/ul/li[1]/div[1]/div[1]/a')
pretended_company.click()
time.sleep(8)

#6th:

    #Now lets save the information of the company:
table_info = driver.find_elements_by_xpath('//*[@id="company_profile_snapshot"]/div/div[2]/span[1]/span')
address = table_info[1].text
phone_num = table_info[2].text
website = table_info[3].text
n_employees= table_info[4].text
sales_vol= table_info[5].text
print(address,phone_num,website,n_employees,sales_vol)

def db_hoovers_info(row):
    
    ##Lets create empty dataframe columns:
    row["company_address"] = np.nan
    row["company_phone_num"] = np.nan
    row["company_website"] = np.nan
    row["company_workers"] = np.nan
    row["sales_volume"] = np.nan
    list_uncaptured = []
    
    for i in range(len(row)):
    ##Step 0:

        driver = webdriver.Chrome(executable_path=r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver', chrome_options=options)
        driver = webdriver.Chrome(executable_path = driver_path)

        driver.get("https://www.dnb.com/solutions/small-business/small-business-strategic-planning.html")

    ##1st:
    #Open the website:
        #driver.get('"https://www.dnb.com/solutions/small-business/small-business-strategic-planning.html"')
        # finding the element that says "Pesquisa Fonética (para Nomes de Estabelecimentos, Insígnias e Logotipos)":
        time.sleep(3)
        filter_button = driver.find_element_by_xpath('//*[@id="search_button"]')
        filter_button.click() ## with the method click you will be able to click the element you searched for ##
        print("got here 1")

    ##2nd:
        #Find the top search area:
        search = driver.find_element_by_xpath('//*[@id="search_button"]')
        search.click()    
        print("got here 2")
    ##3rd:
        #Go to the search area and put the string we want:
        find = driver.find_element_by_xpath('//*[@id="query_display"]')
        time.sleep(1.5)
        find.click()
        print("got here 3")
        find = driver.find_element_by_xpath('//*[@id="search_query"]')
        find.clear()
        time.sleep(1.5)
        print("got here 4")
        try:
            if row["razao_social"].iloc[i] != "nan":
                find.send_keys(row["razao_social"].iloc[i])    
                find.send_keys(Keys.ENTER)
                print("got here 5")
            #5th:
                #Click on the first result for the companies list and wait a bit for the page to load:
                pretended_company =driver.find_element_by_xpath('//*[@id="company_results"]/ul/li[1]/div[1]/div[1]/a')
                pretended_company.click()
                print("got here 6")
                time.sleep(8)

            #6th:
                #Now lets save the information of the company:
                table_info = driver.find_elements_by_xpath('//*[@id="company_profile_snapshot"]/div/div[2]/span[1]/span')
                print("got here 7")
                row["company_address"] = table_info[1].text
                row["company_phone_num"] = table_info[2].text
                row["company_website"] = table_info[3].text
                row["company_workers"] = table_info[4].text
                row["sales_volume"] = table_info[5].text
            else:
                continue
        
        except:
            print("didn't capture",row["razao_social"].iloc[i])
            list_uncaptured += row["razao_social"].iloc[i]
            
    return row

def db_hoovers_use_this(row):
    
    ##Lets create empty dataframe columns:
    row["company_address"] = np.nan
    row["company_phone_num"] = np.nan
    row["company_website"] = np.nan
    row["company_workers"] = np.nan
    row["sales_volume"] = np.nan
    row["extra_column"]= np.nan
    list_uncaptured = []
    row["razao_social"] = row["razao_social"].astype(str)
    for i in range(len(row)):
    ##Step 0:
        if row["razao_social"].iloc[i] != "nan":
        #driver = webdriver.Chrome(executable_path=r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver', chrome_options=options)
            driver_path = r'C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\Selenium_mycode\chromedriver' 

            driver = webdriver.Chrome(executable_path = driver_path)


            driver.get("https://www.dnb.com/solutions/small-business/small-business-strategic-planning.html")

        ##1st:
        #Open the website:
            #driver.get('"https://www.dnb.com/solutions/small-business/small-business-strategic-planning.html"')
            # finding the element that says "Pesquisa Fonética (para Nomes de Estabelecimentos, Insígnias e Logotipos)":
    #         time.sleep(3)
    #         filter_button = driver.find_element_by_xpath('//*[@id="search_button"]')
    #         filter_button.click() ## with the method click you will be able to click the element you searched for ##
            print("got here 1")

        ##2nd:
            #Find the top search area:
            search = driver.find_element_by_xpath('//*[@id="search_button"]')
            search.click()    
            print("got here 2")
        ##3rd:
                    #Go to the search area and put the string we want:
            find = driver.find_element_by_xpath('//*[@id="search_query"]')
            time.sleep(1.5)
            find.click()
            find = driver.find_element_by_xpath('//*[@id="search_query"]')
            find.clear()
            time.sleep(1.5)
            try:


                #find.send_keys(row["razao_social"].iloc[i])
                find.send_keys(row["razao_social"].iloc[i])
                time.sleep(3)
                print("got here 4")
                #Click first business:
                cl = driver.find_element_by_class_name('primary_name')
                cl.click()
                print("got here 5")
                time.sleep(3)
            #5th:
                #Click on the first result for the companies list and wait a bit for the page to load:
#                 pretended_company =driver.find_element_by_xpath('//*[@id="company_results"]/ul/li[1]/div[1]/div[1]/a')
#                 pretended_company.click()
#                 print("got here 6")
#                 time.sleep(8)

                #6th:
                    #Now lets save the information of the company:
                table_info = driver.find_elements_by_xpath('//*[@id="company_profile_snapshot"]/div/div[2]/span[1]/span')
                time.sleep(2)
                row["company_address"].iloc[i] = table_info[1].text
                print(6)
                row["company_phone_num"].iloc[i] = table_info[2].text
                print(7)
                row["company_website"].iloc[i] = table_info[3].text
                print(8)
                row["company_workers"].iloc[i] = table_info[4].text
                print(9)
                row["sales_volume"].iloc[i] = table_info[5].text
                print(10)
                row["extra_column"].iloc[i] = table_info[6].text


#                 //*[@id="company_profile_snapshot"]/div[4]/div[2]/span/span
#                 //*[@id="company_profile_snapshot"]/div[3]/div[2]/span[1]/span
#                 //*[@id="company_profile_snapshot"]/div[2]/div[2]/span/span
            except:
                print("didn't capture",row["razao_social"].iloc[i])
                list_uncaptured += row["razao_social"].iloc[i]
                
        else:
            continue
            
        driver.close()
    return row

db_set = db_hoovers_use_this(data)

db_set.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_050.csv")

time.sleep(2)
data50_100  =pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\facebook50_100.csv")
#type(data["razao_social"].iloc[0])
db_set50_100 = db_hoovers_use_this(data50_100)
db_set50_100.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_50_100.csv")

data50_100["razao_social"] = data50_100["razao_social"].astype(str)
type(data50_100["razao_social"].iloc[1])

data50_100["razao_social"].iloc[1] == "nan"

time.sleep(90)
data100_132  =pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\facebook100_132.csv")
#type(data["razao_social"].iloc[0])
db_set100_132 = db_hoovers_use_this(data100_132)
db_set100_132.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_100_132.csv")


time.sleep(90)
data132_200  =pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\facebook132_200.csv")
#type(data["razao_social"].iloc[0])
db_set132_200 = db_hoovers_use_this(data132_200)
db_set132_200.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_132_200.csv")

time.sleep(90)
data200_250  =pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\facebook200_250.csv")
#type(data["razao_social"].iloc[0])
db_set200_250 = db_hoovers_use_this(data200_250)
db_set200_250.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_200_250.csv")
