==> Step05_cleaning_facebook.ipynb <==
import pandas as pd
import numpy as np
pd.options.mode.chained_assignment = None  # default='warn'
import re
import os

#Import the facebook_dataset:

fb_dset = pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_200_250.csv")
#fb_dset2 = pd.read_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\final_050.csv")
fb_dset.head(5)

#Treating Email column:

fb_dset["email"].fillna(value=0,  inplace=True)
fb_dset["email"].astype(object)
for i in range(len(fb_dset)):
    if fb_dset["email"].iloc[i] == '[]':
        fb_dset["email"].iloc[i] = 0
    else:
        fb_dset["email"].iloc[i] =fb_dset["email"].iloc[i]
        
for i in range(len(fb_dset)):
    if fb_dset["email"].iloc[i] != 0 :
        a=re.findall("[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,3}", fb_dset["email"].iloc[i])
        fb_dset["email"].iloc[i] = a[0]
    
    else:
        fb_dset["email"].iloc[i] = fb_dset["email"].iloc[i] 

fb_dset["email"].head(4)

#Treating Phone number column:

fb_dset["fb_phone_num"].fillna(value=0,  inplace=True)
fb_dset["fb_phone_num"].astype(object)
for i in range(len(fb_dset)):
    if fb_dset["fb_phone_num"].iloc[i] == '[]':
        fb_dset["fb_phone_num"].iloc[i] = 0
    else:
        fb_dset["fb_phone_num"].iloc[i] =fb_dset["fb_phone_num"].iloc[i]
        
for i in range(len(fb_dset)):
    if fb_dset["fb_phone_num"].iloc[i] != 0 :
        b = re.findall("((\d{3} ?){3}|(\d{2} ?\d{3} ?\d{4}?))",fb_dset["fb_phone_num"].iloc[i] )
        fb_dset["fb_phone_num"].iloc[i] = b[0][0].replace(" ","")
    
    else:
        fb_dset["fb_phone_num"].iloc[i] = fb_dset["fb_phone_num"].iloc[i] 

#Deleting numbers not starting in 9 or 2:

for i in range(len(fb_dset)):
    if fb_dset["fb_phone_num"].iloc[i] != 0 :
        if (fb_dset["fb_phone_num"].iloc[i][0] =="9") | (fb_dset["fb_phone_num"].iloc[i][0] =="2" ) :
            fb_dset["fb_phone_num"].iloc[i] = fb_dset["fb_phone_num"].iloc[i] 
        
        else:
            fb_dset["fb_phone_num"].iloc[i] = 0
        
    else:
        fb_dset["fb_phone_num"].iloc[i] = 0
        
fb_dset["fb_phone_num"].head(10)

#Treating Website column, fill with 0 and replace empty lists:

fb_dset["website_fb"].fillna(value=0,  inplace=True)
fb_dset["website_fb"].astype(object)
for i in range(len(fb_dset)):
    if fb_dset["website_fb"].iloc[i] == '[]':
        fb_dset["website_fb"].iloc[i] = 0
    else:
        fb_dset["website_fb"].iloc[i] =fb_dset["website_fb"].iloc[i]

#Treating Website column, replace unwanted [] and '':

for i in range(len(fb_dset)):
    if fb_dset["website_fb"].iloc[i] != 0 :

        fb_dset["website_fb"].iloc[i] = fb_dset["website_fb"].iloc[i].replace("['","").replace("']","")
                
    else:
        fb_dset["website_fb"].iloc[i] = 0
        

fb_dset.tail(3)

fb_dset["company_phone_num"].fillna(value=" ",  inplace=True)

fb_dset["company_website"].fillna(value=" ",  inplace=True)

fb_dset["company_workers"].fillna(value=" ",  inplace=True)

fb_dset["sales_volume"].fillna(value=" ",  inplace=True)



fb_dset['extra_column'] = fb_dset[['company_phone_num', 'company_website', "company_workers","sales_volume"]].agg('_'.join, axis=1)

fb_dset.drop(columns = {"Unnamed: 0","Unnamed: 0.1","Unnamed: 0.1.1","Unnamed: 0.1.1.1","Compl_Name","company_website","company_workers","sales_volume"}, inplace = True)
fb_dset.drop(columns = {"mandatario","morada","codigo_postal","company_address"}, inplace = True)


fb_dset["num_workers"] = np.nan 
for i in range(len(fb_dset)):
    fb_dset["num_workers"].iloc[i] = re.findall("_.*_\$",fb_dset["extra_column"].iloc[i] )
    if fb_dset["num_workers"].iloc[i] == []:
        fb_dset["num_workers"].iloc[i] = re.findall(".*_\$",fb_dset["extra_column"].iloc[i] )
    else:
        fb_dset["num_workers"].iloc[i] = fb_dset["num_workers"].iloc[i]
    
fb_dset["num_workers"]


##Arrange num_workers column:

fb_dset["num_workers"] = np.nan

for i in range(len(fb_dset)):
    if fb_dset["extra_column"].iloc[i] != ' _ _ _ '  :
#         for element in fb_dset["num_workers"].iloc[i]:

        fb_dset["num_workers"].iloc[i] =  fb_dset["extra_column"].iloc[i].replace("www","").replace(",","")
        fb_dset["num_workers"].iloc[i] = re.findall("_?\w+_\$",fb_dset["extra_column"].iloc[i] )

    else:
        fb_dset["num_workers"].iloc[i] = 0


for i in range(len(fb_dset)):
    if fb_dset["num_workers"].iloc[i] == 0  :
        fb_dset["num_workers"].iloc[i] = []

    else :
        fb_dset["num_workers"].iloc[i] = fb_dset["num_workers"].iloc[i]
        
for i in range(len(fb_dset)):
    if fb_dset["num_workers"].iloc[i] != []:
        for element in fb_dset["num_workers"].iloc[i]: 
            fb_dset["num_workers"].iloc[i] = re.findall("\d{1,6}_\$",element)
        
    else:
        fb_dset["num_workers"].iloc[i] = fb_dset["num_workers"].iloc[i]
           
            
#             #re.findall("\d{1,6}_\$",fb_dset["num_workers"].iloc[1][0])
            
for i in range(len(fb_dset)):
    if fb_dset["num_workers"].iloc[i] != []:
        
        for element in fb_dset["num_workers"].iloc[i]:
            fb_dset["num_workers"].iloc[i] = element.replace("_$","")
            
    else:
        fb_dset["num_workers"].iloc[i] = 0
fb_dset["num_workers"]

# ##Arrange num_workers column:

fb_dset["sales_vol"] = np.nan

for i in range(len(fb_dset)):
    if fb_dset["extra_column"].iloc[i] != ' _ _ _ '  :
#         for element in fb_dset["num_workers"].iloc[i]:

        fb_dset["sales_vol"].iloc[i] =  fb_dset["extra_column"].iloc[i].replace("www","").replace(",","").replace(".","").replace("million","0000").replace(" ","").replace("billion","0000")
        fb_dset["sales_vol"].iloc[i] = re.findall("_\$.*_?",fb_dset["sales_vol"].iloc[i] )

    else:
#         fb_dset["sales_vol"].iloc[i] = 0
#         fb_dset["sales_vol"].iloc[i] = int(fb_dset["sales_vol"].iloc[i]
        fb_dset["sales_vol"].iloc[i] = 0



for i in range(len(fb_dset)):
    if fb_dset["sales_vol"].iloc[i] != 0:
        for element in fb_dset["sales_vol"].iloc[i]: 
            fb_dset["sales_vol"].iloc[i] = re.findall("_\$.{1,16}_?",element)
        
    else:
        fb_dset["sales_vol"].iloc[i] = fb_dset["sales_vol"].iloc[i]


    
            
for i in range(len(fb_dset)):
    if fb_dset["sales_vol"].iloc[i] == [] :
        fb_dset["sales_vol"].iloc[i]= 0
    
    elif fb_dset["sales_vol"].iloc[i] != 0:
        for element in fb_dset["sales_vol"].iloc[i]:
            fb_dset["sales_vol"].iloc[i]= element + "_"  

    else:
        fb_dset["sales_vol"].iloc[i]= 0

        
for i in range(len(fb_dset)):
    if fb_dset["sales_vol"].iloc[i] != 0:

        fb_dset["sales_vol"].iloc[i] = re.findall("_\$\d+_",fb_dset["sales_vol"].iloc[i])
        for element in fb_dset["sales_vol"].iloc[i]:
            fb_dset["sales_vol"].iloc[i]= element 
        fb_dset["sales_vol"].iloc[i] = fb_dset["sales_vol"].iloc[i].replace("_","").replace("$","")
    else:
        fb_dset["sales_vol"].iloc[i] = fb_dset["sales_vol"].iloc[i]
           
fb_dset["num_workers"] =fb_dset["num_workers"].astype(int)
fb_dset["sales_vol"] = fb_dset["sales_vol"].astype(int)
            

fb_dset["sales_vol"]

##Save dataframe:

fb_dset.to_csv(r"C:\Users\crocs\OneDrive - Universidade de Lisboa\Ambiente de Trabalho\Gonçalo\College&Courses\Data Analytics- IronHack\Course Data Analytics\Projects\FinalProject\Mycode\database\last\last_200_250.csv")





#Columns to get to the Algorithm:

    #Rating , user_rating_total , got website ,got facebook , following_fb , num_workers , sales_vol
